refactor(ggtfs): agency handling
This commit brings in a new way to parse CSV rows, while removing pointers from the returned agency structs (this has some side effects in the other structs, which will be handled properly in later commits). The original idea was to use pointers on optional values, so their absence could be checked with nil value, but this seems more of a validation issue than data type issue.