fix: implement cache controls
This commit implements cache controls which allow the cached items to expire according to configured rules. Two cache modes are supported: short and long. The short cache can be used for example during night when the service day is about to expire, and the cache must invalidate during the period between last journey of the previous service day and the start of the next service day.